UCT/MD: add memory access flags: cr fixes p1
