+++ src/uct/base/uct_md.c
@@ -472,13 +472,11 @@ ucs_status_t uct_md_query(uct_md_h md, uct_md_attr_t *md_attr)
 
 static ucs_status_t uct_mem_check_flags(unsigned flags)
 {
-    if (!(flags & UCT_MD_MEM_ACCESS_DEFAULT)) {
-        ucs_warn("no memory access flags set");
+    if (!(flags & UCT_MD_MEM_ACCESS_ALL)) {
         return UCS_ERR_INVALID_PARAM;
     }
     return UCS_OK;
 }
-
 
 ucs_status_t uct_md_mem_alloc(uct_md_h md, size_t *length_p, void **address_p,
                               unsigned flags, const char *alloc_name, uct_mem_h *memh_p)
@@ -489,6 +487,7 @@ ucs_status_t uct_md_mem_alloc(uct_md_h md, size_t *length_p, void **address_p,
     if (status != UCS_OK) {
         return status;
     }
+
     return md->ops->mem_alloc(md, length_p, address_p, flags, memh_p UCS_MEMTRACK_VAL);
 }
