+++ test/gtest/uct/test_md.h
@@ -1,0 +1,41 @@
+/**
+* Copyright (C) Mellanox Technologies Ltd. 2001-2017.  ALL RIGHTS RESERVED.
+*
+* See file LICENSE for terms.
+*/
+
+
+#ifndef UCT_TEST_MD
+#define UCT_TEST_MD
+
+class test_md : public testing::TestWithParam<std::string>,
+                public ucs::test_base
+{
+public:
+    UCS_TEST_BASE_IMPL;
+
+    static std::vector<std::string> enum_mds(const std::string& mdc_name);
+
+    test_md();
+
+protected:
+    virtual void init();
+    virtual void cleanup();
+    virtual void modify_config(const std::string& name, const std::string& value,
+                               bool optional);
+    void check_caps(uint64_t flags, const std::string& name);
+
+    void test_registration();
+
+    uct_md_h pd() {
+        return m_pd;
+    }
+
+    static void* alloc_thread(void *arg);
+
+private:
+    ucs::handle<uct_md_config_t*> m_md_config;
+    ucs::handle<uct_md_h>         m_pd;
+};
+
+#endif
